Revert "- feature: added button to sync ollama models with agents (#208)"
This reverts commit f9fde517e24e1ecc1975c68322907e42f53aafd6.
