--- .gitattributes
@@ -1,2 +1,0 @@
-apps/shinkai-desktop/src-tauri/bin/* filter=lfs diff=lfs merge=lfs -text
-apps/shinkai-desktop/src-tauri/bin/** filter=lfs diff=lfs merge=lfs -text
