+++ .gitignore
@@ -44,4 +44,3 @@ apps/shinkai-visor-e2e/src/shinkai-node/qr_code_device_0.png
 qr_code_device_0.png
 apps/shinkai-visor-e2e/test-report
 apps/shinkai-desktop/src-tauri/storage
-apps/shinkai-desktop/src-tauri/node_storage
