--- apps/shinkai-desktop/src-tauri/.taurignore
@@ -1,1 +1,0 @@
-node_storage/
