+++ apps/shinkai-desktop/src-tauri/Cargo.toml
@@ -21,12 +21,9 @@ cpal = { git = "https://github.com/RustAudio/cpal.git", features = ["asio"] }
 webrtc-vad = "0.2"
 hound = "3.4.0"
 # fix this dependency later on
-reqwest = { version = "0.11", features = ["json", "stream"] }
+reqwest = { version = "0.11", features = ["json"] }
 lazy_static = "1.4.0"
 tokio = "1.36.0"
-chrono = "0.4.38"
-futures-util = "0.3"
-regex = "1.10.4"
 
 [features]
 # this feature is used for production builds or when `devPath` points to the filesystem
