--- apps/shinkai-desktop/src-tauri/bin/ollama-v0.1.33-aarch64-apple-darwin
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e77993c760eadf226feffdda8af1736e21f2edbfcce134fb6808f328ae8dd371
-size 52039248
