--- apps/shinkai-desktop/src-tauri/bin/ollama-v0.1.33-x86_64-pc-windows-msvc.exe
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:0379c8434a034f18ed176361457c7d960b33d354c242f1535a5b265002cbd6fd
-size 52393408
