--- apps/shinkai-desktop/src-tauri/bin/ollama-v0.1.33-x86_64-unknown-linux-gnu
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:c7416d3051fc50606a2151cc5072bff8c77b9608f8bd7f4f489ce6de454ecfef
-size 305057208
