--- apps/shinkai-desktop/src-tauri/bin/shinkai-node-v0.6.5-aarch64-apple-darwin
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:d8d023050ef7906ef099623b251e85375edf7088c1e445ba63749d0bf722b2dc
-size 49189200
