--- apps/shinkai-desktop/src-tauri/bin/shinkai-node-v0.6.5-x86_64-pc-windows-msvc.exe
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:459a136b67e162b771ee4fd29237d1150adb62e6ac5610fe47d534ac77af55bd
-size 35965952
