--- apps/shinkai-desktop/src-tauri/bin/shinkai-node-v0.6.5-x86_64-unknown-linux-gnu
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:fb32a9ac530d5d58d0883f4f3962dc6ddb7701125239d941df94043954921953
-size 52249552
