+++ apps/shinkai-desktop/src-tauri/src/local_shinkai_node/mod.rs
@@ -1,5 +1,3 @@
-pub mod ollama_api;
-pub mod process_handlers;
 pub mod shinkai_node_manager;
-pub mod shinkai_node_manager_instance;
 pub mod shinkai_node_options;
+pub mod shinkai_node_manager_instance;
