--- apps/shinkai-desktop/src-tauri/src/local_shinkai_node/ollama_api/mod.rs
@@ -1,2 +1,0 @@
-pub mod ollama_api_client;
-pub mod ollama_api_types;
