--- apps/shinkai-desktop/src-tauri/src/local_shinkai_node/process_handlers/mod.rs
@@ -1,5 +1,0 @@
-pub mod process_handler;
-pub mod process_utils;
-pub mod ollama_process_handler;
-pub mod shinkai_node_process_handler;
-pub mod logger;
