+++ apps/shinkai-desktop/src-tauri/src/local_shinkai_node/shinkai_node_manager_instance.rs
@@ -1,9 +1,8 @@
 use std::sync::Arc;
 
-use crate::local_shinkai_node::shinkai_node_manager::ShinkaiNodeManager;
 use lazy_static::lazy_static;
 use tokio::sync::Mutex;
+use crate::local_shinkai_node::shinkai_node_manager::ShinkaiNodeManager;
 lazy_static! {
-    pub static ref SHINKAI_NODE_MANAGER_INSTANCE: Arc<Mutex<ShinkaiNodeManager>> =
-        Arc::new(Mutex::new(ShinkaiNodeManager::new()));
+    pub static ref SHINKAI_NODE_MANAGER_INSTANCE: Arc<Mutex<ShinkaiNodeManager>> = Arc::new(Mutex::new(ShinkaiNodeManager::new()));
 }
