+++ apps/shinkai-desktop/src/windows/shinkai-node-manager/main.tsx
@@ -2,9 +2,6 @@ import './globals.css';
 
 import { zodResolver } from '@hookform/resolvers/zod';
 import {
-useSyncOllamaModels
-} from '@shinkai_network/shinkai-node-state/lib/mutations/syncOllamaModels/useSyncOllamaModels';
-import {
   AlertDialog,
   AlertDialogAction,
   AlertDialogCancel,
@@ -27,7 +24,6 @@ import {
 } from '@shinkai_network/shinkai-ui';
 import { QueryClientProvider } from '@tanstack/react-query';
 import {
-  Bot,
   ListRestart,
   Loader2,
   PlayCircle,
@@ -69,7 +65,6 @@ import {
 initSyncStorage();
 
 const App = () => {
-  const auth = useAuth(auth => auth.auth);
   const setLogout = useAuth((auth) => auth.setLogout);
   const { setShinkaiNodeOptions } = useShinkaiNodeManager();
   const logsScrollRef = useRef<HTMLDivElement | null>(null);
@@ -141,8 +136,6 @@ const App = () => {
   const shinkaiNodeOptionsFormWatch = useWatch({
     control: shinkaiNodeOptionsForm.control,
   });
-  const { mutateAsync: syncOllamaModels, isPending: syncOllamaModelsIsPending } =
-    useSyncOllamaModels({});
 
   useEffect(() => {
     logsScrollRef.current?.scrollIntoView({
@@ -160,19 +153,6 @@ const App = () => {
     shinkaiNodeRemoveStorage();
   };
 
-  const startSyncOllamaModels = (): void => {
-    syncOllamaModels({
-      nodeAddress: auth?.node_address ?? '',
-      senderSubidentity: auth?.profile ?? '',
-      shinkaiIdentity: auth?.shinkai_identity ?? '',
-      sender: auth?.node_address ?? '',
-      my_device_encryption_sk: auth?.my_device_encryption_sk ?? '',
-      my_device_identity_sk: auth?.my_device_identity_sk ?? '',
-      node_encryption_pk: auth?.node_encryption_pk ?? '',
-      profile_encryption_sk: auth?.profile_encryption_sk ?? '',
-      profile_identity_sk: auth?.profile_identity_sk ?? '',
-     });
-  }
   return (
     <div className="h-full w-full overflow-hidden p-8">
       <div className="flex flex-row items-center">
@@ -207,7 +187,7 @@ const App = () => {
             onClick={() => shinkaiNodeKill()}
             variant={'default'}
           >
-            {shinkaiNodeKillIsPending ? (
+            {shinkaiNodeSpawnIsPending || shinkaiNodeKillIsPending ? (
               <Loader2 className="animate-spin" />
             ) : (
               <StopCircle className="" />
@@ -223,18 +203,6 @@ const App = () => {
               <Loader2 className="animate-spin" />
             ) : (
               <Trash className="" />
-            )}
-          </Button>
-
-          <Button
-            disabled={!shinkaiNodeIsRunning}
-            onClick={() => startSyncOllamaModels()}
-            variant={'default'}
-          >
-            {syncOllamaModelsIsPending ? (
-              <Loader2 className="animate-spin" />
-            ) : (
-              <Bot className="" />
             )}
           </Button>
         </div>
@@ -257,8 +225,7 @@ const App = () => {
                     return (
                       <>
                         <div className="text-gray-80 text-sm" key={index}>
-                          {'ℹ️'} {new Date(log.timestamp * 1000).toISOString()}{' '}
-                          | {log.process} | {log.message}
+                          {'>'} {log}
                         </div>
                         <Separator className="my-2" />
                       </>
