+++ apps/shinkai-desktop/src/windows/shinkai-node-manager/shinkai-node-process-client.ts
@@ -20,12 +20,6 @@ export type ShinkaiNodeOptions = {
   starting_num_qr_devices?: number;
 };
 
-export type LogEntry = {
-  timestamp: number;
-  process: string;
-  message: string;
-};
-
 // Client
 export const queryClient = new QueryClient();
 
@@ -43,14 +37,14 @@ export const useShinkaiNodeIsRunningQuery = (
 export const useShinkaiNodeGetLastNLogsQuery = (
   input: { length: number },
   options?: QueryObserverOptions,
-): UseQueryResult<LogEntry[], Error> => {
+): UseQueryResult<string[], Error> => {
   const query = useQuery({
     queryKey: ['shinkai_node_get_last_n_logs'],
-    queryFn: (): Promise<LogEntry[]> =>
+    queryFn: (): Promise<string[]> =>
       invoke('shinkai_node_get_last_n_logs', { length: input.length }),
     ...options,
   });
-  return { ...query } as UseQueryResult<LogEntry[], Error>;
+  return { ...query } as UseQueryResult<string[], Error>;
 };
 export const useShinkaiNodeGetOptionsQuery = (
   options?: QueryObserverOptions,
