+++ libs/shinkai-message-ts/src/wasm/ShinkaiMessageBuilderWrapper.ts
@@ -1357,69 +1357,4 @@ export class ShinkaiMessageBuilderWrapper {
     const message = builder.build_to_string();
     return message;
   }
-
-  static scanOllamaModels(
-    my_encryption_secret_key: string,
-    my_signature_secret_key: string,
-    receiver_public_key: string,
-    sender: string,
-    sender_subidentity: string,
-    receiver: string,
-    receiver_subidentity: string,
-  ): string {
-    const builder = new ShinkaiMessageBuilderWrapper(
-      my_encryption_secret_key,
-      my_signature_secret_key,
-      receiver_public_key,
-    );
-
-    builder.message_raw_content('');
-    builder.message_schema_type(
-      MessageSchemaType.APIScanOllamaModels.toString(),
-    );
-    builder.internal_metadata(
-      sender_subidentity,
-      receiver_subidentity,
-      '',
-      'None',
-    );
-    builder.external_metadata_with_intra(receiver, sender, sender_subidentity);
-    builder.body_encryption('DiffieHellmanChaChaPoly1305');
-
-    const message = builder.build_to_string();
-    return message;
-  }
-
-  static addOllamaModels(
-    my_encryption_secret_key: string,
-    my_signature_secret_key: string,
-    receiver_public_key: string,
-    sender: string,
-    sender_subidentity: string,
-    receiver: string,
-    receiver_subidentity: string,
-    payload: { models: string[] },
-  ): string {
-    const builder = new ShinkaiMessageBuilderWrapper(
-      my_encryption_secret_key,
-      my_signature_secret_key,
-      receiver_public_key,
-    );
-
-    builder.message_raw_content(JSON.stringify(payload));
-    builder.message_schema_type(
-      MessageSchemaType.APIAddOllamaModels.toString(),
-    );
-    builder.internal_metadata(
-      sender_subidentity,
-      receiver_subidentity,
-      '',
-      'None',
-    );
-    builder.external_metadata_with_intra(receiver, sender, sender_subidentity);
-    builder.body_encryption('DiffieHellmanChaChaPoly1305');
-
-    const message = builder.build_to_string();
-    return message;
-  }
 }
