+++ libs/shinkai-node-state/src/lib/constants.ts
@@ -16,5 +16,4 @@ export enum FunctionKey {
   GET_AVAILABLE_SHARED_ITEMS = 'GET_AVAILABLE_SHARED_ITEMS',
   GET_MY_SUBSCRIPTIONS = 'GET_MY_SUBSCRIPTIONS',
   GET_MY_SHARED_FOLDERS = 'GET_MY_SHARED_FOLDERS',
-  SCAN_OLLAMA_MODELS = 'SCAN_OLLAMA_MODELS',
 }
